break spotlight history and z/osmf apis into separate services
