+++ virtual-desktop/pluginDefinition.json
@@ -16,6 +16,9 @@
       "actions": {
         "aggregationPolicy": "none"
       },
+      "spotlight": {
+        "aggregationPolicy": "none"
+      },
       "ui": {
         "subResources":{
           "launchbar": {
